pyproject.toml
[project]
name = "lexguard"
version = "0.1.0"
description = "Compliance reasoning engine"
requires-python = ">=3.10"

dependencies = [
    "pydantic",
    "pymupdf",
    "pdfplumber",
    "tqdm",
    "rank-bm25",
    "rich",
    "sentence-transformers",
    "faiss-cpu",
    "numpy"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"